Finish refactor; implement payment-making
