+++ src/client/index.ts
@@ -1,2 +1,5 @@
 export * from './payMcpClient.js';
+export * from './oAuth.js'
+export * from './payMcpFetcher.js'
+export * from './solanaPaymentMaker.js'
 export * from './types.js';
