+++ src/client/oauth.test.ts
@@ -1,9 +1,10 @@
-import { SqliteOAuthDb } from './oAuthDb';
-import { OAuthClient, OAuthAuthenticationRequiredError } from './oAuth.js';
+import { SqliteOAuthDb } from '../common/oAuthDb';
+import { OAuthClient } from './oAuth.js';
 import { describe, it, expect } from 'vitest';
 import fetchMock from 'fetch-mock';
-import { FetchLike, OAuthDb } from './types.js';
-import { mockResourceServer, mockAuthorizationServer } from './testHelpers.js';
+import { FetchLike, OAuthDb } from '../common/types.js';
+import { mockResourceServer, mockAuthorizationServer } from '../testHelpers.js';
+import { OAuthAuthenticationRequiredError } from './oAuth.js';
 
 function oauthClient(fetchFn: FetchLike, db?: OAuthDb, isPublic: boolean = false, strict: boolean = true, callbackUrl: string = 'https://example.com/mcp/callback') {
   return new OAuthClient({
