+++ src/client/platform/bundlerCompatibility.test.ts
