+++ src/client/solanaPaymentMaker.ts
@@ -1,11 +1,9 @@
 import type { PaymentMaker } from './types.js';
 import { Keypair, Connection, PublicKey, ComputeBudgetProgram, sendAndConfirmTransaction } from "@solana/web3.js";
 import { createTransfer, ValidateTransferError as _ValidateTransferError } from "@solana/pay";
-import nacl from "tweetnacl";
-import naclUtil from "tweetnacl-util";
 import bs58 from "bs58";
 import BigNumber from "bignumber.js";
-import { generateJWT } from './jwt.js';
+import { generateJWT } from '../common/jwt.js';
 import { importJWK } from 'jose';
 
 // this is a global public key for USDC on the solana mainnet
@@ -28,7 +26,7 @@ export class SolanaPaymentMaker implements PaymentMaker {
     this.source = Keypair.fromSecretKey(bs58.decode(sourceSecretKey));
   }
   
-  generateJWT = async(paymentIds?: string[]): Promise<string> => {
+  generateJWT = async({paymentIds, codeChallenge}: {paymentIds?: string[], codeChallenge?: string}): Promise<string> => {
     // Solana/Web3.js secretKey is 64 bytes: 
     // first 32 bytes are the private scalar, last 32 are the public key.
     // JWK expects only the 32-byte private scalar for 'd'
@@ -39,7 +37,7 @@ export class SolanaPaymentMaker implements PaymentMaker {
       x: Buffer.from(this.source.publicKey.toBytes()).toString('base64url'),
     };
     const privateKey = await importJWK(jwk, 'EdDSA');
-    return generateJWT(this.source.publicKey.toBase58(), privateKey, paymentIds);
+    return generateJWT(this.source.publicKey.toBase58(), privateKey, paymentIds || [], codeChallenge || '');
   }
 
   makePayment = async (amount: BigNumber, currency: string, receiver: string): Promise<string> => {
