+++ src/index.ts
@@ -4,17 +4,6 @@ if (typeof navigator !== 'undefined' && navigator.product === 'ReactNative') {
   require('react-native-url-polyfill/auto');
 }
 
-export * from './auth.js';
-export * from './oAuth.js';
-export * from './oAuthDb.js';
-export * from './oAuthResource.js';
-export * from './payMcpClient.js';
-export * from './solanaPaymentMaker.js';
-export * from './types.js';
-export * from './server/index.js';
-export * from './server/types.js';
-export * from './common/logger.js';
+export * from './common/index.js';
 export * from './server/index.js';
 export * from './client/index.js';
-export * from './client/types.js';
-export * from './common/types.js';
