+++ src/server/getResource.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { getProtectedResourceMetadata } from './protectedResourceMetadata.js';
-import * as TH from './testHelpers.js';
+import * as TH from './serverTestHelpers.js';
 import { getResource } from './getResource.js';
 
 describe('getResource', () => {
