+++ src/server/index.ts
@@ -1,4 +1,4 @@
-export { payMcpServer } from './payMcpServer.js';
-export { requirePayment } from './requirePayment.js';
+export * from './payMcpServer.js';
 export * from './types.js';
-export { payMcpUser } from './payMcpContext.js';
+export * from './payMcpContext.js';
+export * from './requirePayment.js';
