+++ src/server/oAuthChallenge.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import * as TH from './testHelpers.js';
+import * as TH from './serverTestHelpers.js';
 import { TokenProblem } from './types.js';
 import { sendOAuthChallenge } from './oAuthChallenge.js';
