+++ src/server/paymentServer.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, vi } from 'vitest';
-import * as TH from './testHelpers.js';
+import * as TH from './serverTestHelpers.js';
 import { PayMcpPaymentServer } from './paymentServer.js';
 import { BigNumber } from 'bignumber.js';
 import fetchMock from 'fetch-mock';
-import { SqliteOAuthDb } from '../oAuthDb.js';
+import { SqliteOAuthDb } from '../common/oAuthDb.js';
 
 describe('PayMcpPaymentServer', () => {
   it('should call the charge endpoint', async () => {
@@ -165,8 +165,7 @@ describe('PayMcpPaymentServer', () => {
       ...TH.charge({
         source: 'test-source',
         destination: 'test-destination'
-      }),
-      resource: new URL('https://example.com/resource')
+      })
     });
 
     // Verify the specific credentials were used
@@ -185,8 +184,7 @@ describe('PayMcpPaymentServer', () => {
       ...TH.charge({
         source: 'test-source',
         destination: 'test-destination'
-      }),
-      resource: new URL('https://example.com/resource')
+      })
     })).rejects.toThrow('No client credentials found');
 
     // Verify fetch was never called
@@ -276,8 +274,7 @@ describe('PayMcpPaymentServer', () => {
       ...TH.charge({
         source: 'test-source',
         destination: 'test-destination'
-      }),
-      resource: new URL('https://example.com/resource')
+      })
     })).rejects.toThrow('POST /payment-request responded with unexpected HTTP status 400');
   });
 
@@ -303,8 +300,7 @@ describe('PayMcpPaymentServer', () => {
       ...TH.charge({
         source: 'test-source',
         destination: 'test-destination'
-      }),
-      resource: new URL('https://example.com/resource')
+      })
     })).rejects.toThrow('POST /payment-request response did not contain an id');
   });
 });
