+++ src/server/requirePayment.ts
@@ -1,16 +1,6 @@
-import { RequirePaymentConfig, PayMcpConfig } from "./types.js";
+import { RequirePaymentConfig } from "../common/types.js";
 import { getPayMcpConfig, getPayMcpResource, payMcpUser } from "./payMcpContext.js";
-
-export class PaymentRequestError extends Error { 
-  paymentRequestId: string;
-  paymentRequestUrl: string;
-  constructor(config: PayMcpConfig, paymentRequestId: string) {
-    const url = `${config.server}/payment-request/${paymentRequestId}`;
-    super(`Payment request failed. Please pay at: ${url}`);
-    this.paymentRequestId = paymentRequestId;
-    this.paymentRequestUrl = url;
-  }
-}
+import { PaymentRequestError } from "../common/paymentRequestError.js";
 
 export async function requirePayment(paymentConfig: RequirePaymentConfig): Promise<void> {
   const config = getPayMcpConfig();
@@ -45,10 +35,10 @@ export async function requirePayment(paymentConfig: RequirePaymentConfig): Promi
   const existingPaymentId = await paymentConfig.getExistingPaymentId?.();
   if (existingPaymentId) {
     config.logger.info(`Found existing payment ID ${existingPaymentId}`);
-    throw new PaymentRequestError(config, existingPaymentId)
+    throw new PaymentRequestError(config.server, existingPaymentId)
   }
 
-  const paymentId = await config.paymentServer.createPaymentRequest({...charge, resource});
+  const paymentId = await config.paymentServer.createPaymentRequest(charge);
   config.logger.info(`Created payment request ${paymentId}`);
-  throw new PaymentRequestError(config, paymentId);
+  throw new PaymentRequestError(config.server, paymentId);
 }
