+++ src/types.ts
@@ -1,58 +1,1 @@
 import type { BigNumber } from 'bignumber.js';
-import type { Request } from 'express';
-
-export type OpPriceMap = { [key: string]: number };
-
-export type OpPrices = OpPriceMap | (({request}: {request: Request}) => OpPriceMap);
-
-export type CustomJWTPayload = {
-  paymentIds?: string[];
-}
-
-export interface PaymentMaker {
-  makePayment: (amount: BigNumber, currency: string, receiver: string, resourceName?: string) => Promise<string>;
-  generateJWT: (paymentIds?: string[]) => Promise<string>;
-}
-
-export type ClientCredentials = {
-  clientId: string,
-  clientSecret: string,
-  redirectUri: string
-};
-
-export type PKCEValues = {
-  codeVerifier: string,
-  codeChallenge: string,
-  resourceUrl: string,
-  url: string
-};
-
-export type AccessToken = {
-  accessToken: string,
-  refreshToken?: string,
-  expiresAt?: number,
-  resourceUrl: string
-};
-
-export interface OAuthResourceDb {
-  getClientCredentials(serverUrl: string): Promise<ClientCredentials | null>;
-  saveClientCredentials(serverUrl: string, credentials: ClientCredentials): Promise<void>;
-  close(): Promise<void>;
-}
-
-export interface OAuthDb extends OAuthResourceDb {
-  getPKCEValues(userId: string, state: string): Promise<PKCEValues | null>;
-  savePKCEValues(userId: string, state: string, values: PKCEValues): Promise<void>;
-  getAccessToken(userId: string, url: string): Promise<AccessToken | null>;
-  saveAccessToken(userId: string, url: string, token: AccessToken): Promise<void>;
-}
-
-export type TokenData = {
-  active: boolean,
-  scope?: string,
-  sub?: string,
-  aud?: string|string[],
-}
-
-// This should match MCP SDK's version, however they don't export it
-export type FetchLike = (url: string | URL, init?: RequestInit) => Promise<Response>;
